Creating a base class DeepgramFluxSTTBase so we can reuse Deepgram Flux logic.
